+++ TrainApp/Common/Business/资源/YXWholeDatumFetcher.m
@@ -15,7 +15,7 @@
 
 @implementation YXWholeDatumFetcher
 
-- (void)startWithBlock:(void(^)(int total, NSArray *retItemArray, NSError *error))aCompleteBlock {
+- (void)startWithBlock:(void(^)(NSInteger total, NSArray *retItemArray, NSError *error))aCompleteBlock {
     [self stop];
     self.request = [[YXDatumSearchRequest alloc] init];
     self.request.condition = self.condition;
